# Conflicts:
#	lib/data/remote/apis/api_path.dart
#	lib/l10n/arb/app_en.arb

+++ lib/screens/transfer_balance/bloc/transfer_balance_state.dart
@@ -2,12 +2,12 @@ part of 'transfer_balance_bloc.dart';
 
 
 class TransferBalanceState extends Equatable{
-  final double? amount;
+  final int? amount;
 
   const TransferBalanceState({required this.amount});
 
   TransferBalanceState copyWith(
-      {double? amount,}) {
+      {int? amount,}) {
     return TransferBalanceState(
       amount: amount ?? this.amount,
     );
